requirements.txt
pandas
matplotlib
numpy
notebook
jupyter-lab
mysql-connector-python
Flask
python-dotenv
pyodbc